requirements.txt
streamlit==1.47.0
langchain==0.3.26
langchain-community==0.3.27
langchain-huggingface==0.3.1
langchain-openai==0.3.28
python-dotenv==1.1.1
faiss-cpu==1.11.0.post1
sentence-transformers==5.0.0